src/simulation-engine/requirements.txt
camel-oasis==0.2.5
camel-ai==0.2.78
zep-cloud==3.13.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
